requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
flake8==3.9.1
gunicorn==20.1.0
idna==2.10
mccabe==0.6.1
newrelic==6.2.0.156
ply==3.11
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-http-client==3.3.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stone==3.2.1
urllib3==1.26.4
whitenoise==5.2.0
